Changes not in dev
* Fixes race condition on jobs_.empty() in PreviewLoader::processNextJob()
* Fixes issues related to not waiting for currently running jobs to finish when forcefully ending thumbnail first pass processing

FileBrowser works with individual Gtk filesystem events instead of reparsing the entire directory after every event.

DirBrowser handles delete and create events individually

Fixes issues that came up in the pull request review
- Bug: existence of a wrong key name was being checked from the config file.
- Bug: previewFailed was called twice if the file was missing. The bug can be confirmed by starting a directory copy, cancelling the copy mid way and instantly deleting the partially copied directory.
- Bug: previewFailed didn't update the progress bar.
- Bug: matching the beginnings of paths as a shortcut to find all removed files from the File Browser and directories from directry monitors was failing when there were directories that had the exact beginning part of the path, but the last directory name in the path was longer. I.e. files and directories were being removed erroneously.
- Bug: If the newFileDelayTime was zero and the user renamed a file, the old file was not removed from the File Browser.
- As part of the previous fix, the timer is now disconnected when directory is closed and connected when a directory is opened. This allows users changing the newFileDelayTime value and apply the changes in all cases by double clicking or refreshing any directory. Previously, it was not possible if the value was 0 when program was started. Similarly, disabling the timer completely was not possible without restarting the program.
- Bug: getFilesRecursively was called with wrongly calculated quota.
- In DirBrowser, Glib::RefPtr<Gio::File> was not passed as reference while it was preferable.
- Minor typos and documentation mistakes.

Fixes some issues with getFileList I missed in my fixes for the PR review bugs. Also, the root is checked for missed files even if the quota doesn't allow searching further on the tree. newDirs is never empty because the root is added to it, unless the quota calculation reveals a bug and goes below 0. It's still safe to be called.

The maximum directory quota was not taken into account  resulting in creating excess directory monitors and files being added from directories that weren't supposed to be monitored.